feat(linsolve): add a mixed-precision builtin<float> AMGCL backend, opt-in
Templates the AMGCL wrapper (amgcl-wrapper.hpp) on the backend's value
type, adding a float32 hierarchy (LinearSolverFloat) alongside the
existing double one, selected per PyAMGCLSolver instance via a new
"backendPrecision" option ("double" default, "float" opt-in). The near
null-space stays double unconditionally -- confirmed against AMGCL's own
headers that coarsening::nullspace_params::B is hardcoded
std::vector<double>, independent of the solve backend (a plan-vs-library
mismatch found while implementing, not a design choice). rhs/x at the
applyPreconditioner()/solve() boundary stay double too; the C++ wrapper
converts internally so the hot path (one call per outer Krylov iteration)
never crosses the Cython boundary with a dtype-dependent signature.

Not wired as blockamg's default: validated offline on all 9 dumped ords,
float32 inflated outer GMRES iterations by up to 30% (plan expected
<=10%) and only won on wall-clock roughly half the time, netting ~3%
aggregate rather than the ~1.3-1.5x hoped for. Kept as an opt-in via
fieldPreconds's "backendPrecision" key for problems where it might do
better (plausibly the chebyshev smoother's power-iteration
spectral-radius estimate, already flagged as sensitive in §17 B5, losing
accuracy in float32 on this problem).

+++ edelweissfe/linsolve/blockamg/blockamg.py
@@ -78,17 +78,22 @@ from scipy.sparse.linalg import LinearOperator, gmres
 from edelweissfe.linsolve.base import FieldBlock, FieldStructureAwareLinearSolver
 
 # "backendPrecision" is not an AMGCL parameter -- it selects the AMGCL wrapper's own backend value
-# type (§19.3). __call__ pops it out of whichever precond dict applies (default or a fieldPreconds
-# override) before forwarding the rest verbatim as the AMGCL JSON parameter tree.
+# type (§19.3: "double" (default) or "float"). __call__ pops it out of whichever precond dict applies
+# (default or a fieldPreconds override) before forwarding the rest verbatim as the AMGCL JSON parameter
+# tree. Defaults to "double": measured on all 9 dumped ords, "float" inflated outer GMRES iterations by
+# up to 30% (plan expected <=10%) and only won on wall-clock ~half the time, netting ~3% aggregate, not
+# the ~1.3-1.5x hoped for -- plausibly the chebyshev smoother's power-iteration spectral-radius estimate
+# (§17 B5 already flagged this as sensitive) losing accuracy in float32. Available as an opt-in via
+# fieldPreconds (e.g. {"displacement": {**_DEFAULT_VECTOR_PRECOND, "backendPrecision": "float"}}).
 _DEFAULT_VECTOR_PRECOND = {
-    "backendPrecision": "float",
+    "backendPrecision": "double",
     "coarsening": {"type": "smoothed_aggregation", "aggr": {"eps_strong": 0.01}},
     "relax": {"type": "chebyshev", "degree": 5, "power_iters": 50, "lower": 0.01},
     "npre": 1,
     "npost": 1,
 }
 _DEFAULT_SCALAR_PRECOND = {
-    "backendPrecision": "float",
+    "backendPrecision": "double",
     "coarsening": {"type": "smoothed_aggregation"},
     "relax": {"type": "chebyshev"},
 }
